requirements.txt
pygame>=2.1.0
pdf2image>=1.16.0
Pillow>=8.4.0
